Create updatedocumentation
Angular needs the documentation to be simplified in my opinion. Especially since it’s a Google product, if we want people to begin using Angular as much as React (or JQuery).

Also a prerequisites section should be added to the page, in clear sight with exactly how much HTML5, CSS3, and JavaScript are required to start building stuff in Angular.